Fix up ast_impl
