+++ crates/oxc_regular_expression/src/ast_impl/mod.rs
@@ -1,2 +1,3 @@
 mod display;
+mod span;
 pub mod visit;
